Mention that choosing data structures is possible
I certainly don't want the programming language to require that the
compiler choose the exact data structure that's being used in any given
circumstance.  A developer should be able to do that if he wants.
However, there are many circumstances where that choice is not
particularly important.  In those cases, the developer shouldn't need to
care.