Simplify proposed syntax for public predicate
Instead of requiring each lemma predicate to be named with a trailing
underscore, stipulate that only the first predicate in a file is
public.  The others are lemmas by default.

This eliminates the need for a special syntax error.  It also eliminates
funky naming conventions.

Dart describes privacy with naming conventions.  Although one gets used
to it eventually, it's ugly for new comers.